added function that tells what type of line it is

+++ Interpretter_project/src/interpretter_project/Variable.java
@@ -76,7 +76,7 @@ public class Variable {
             return "int";
         }
         else if(isDouble==true){
-            return "int";
+            return "double";
         }
         else if(isChar==true){
             return "char";
